slate: finish implementing preservation of scroll position properly when editing collaboratively.  Overall this is a really good solution!
